requirements_actor_critic.txt
# Actor-Critic Brush Explorer Requirements
# Install with: pip install -r requirements_actor_critic.txt

# Core dependencies from original DoodleAgent
anthropic>=0.25.0
selenium>=4.15.0
pillow>=10.0.0
python-dotenv>=1.0.0

# Additional dependencies for enhanced functionality
numpy>=1.24.0
matplotlib>=3.7.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For advanced image processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Optional: For data analysis of exploration logs
pandas>=2.0.0
seaborn>=0.12.0 